api/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
python-socketio==5.8.0
python-dotenv==1.0.0
python-engineio==4.5.1
flask-sqlalchemy>=3.0.0
psycopg2-binary>=2.9.0
eventlet==0.33.3 